backend/requirements.txt
fastapi==0.133.1
uvicorn==0.35.0

motor==3.7.0
pymongo==4.12.1

pydantic-settings==2.13.1
livekit-api==0.9.0
protobuf==5.29.3
